pyproject.toml
[tool.poetry]
name = "mitotem-interactive"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "mitotem_interactive"}]

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.0"
opencv-python = "^4.8.1.78"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
robotframework = "^6.1.1"
robotframework-seleniumlibrary = "^6.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
